Silenced warnings.

+++ include/sprout/math/floor.hpp
@@ -68,7 +68,8 @@ namespace sprout {
 				: sprout::math::detail::builtin_floor(x)
 #else
 				: x == 0 ? x
-				: sprout::numeric_limits<std::uintmax_t>::max() < x || sprout::numeric_limits<std::uintmax_t>::max() < -x
+                : (static_cast<FloatType>(sprout::numeric_limits<std::uintmax_t>::max()) <  x ||
+                   static_cast<FloatType>(sprout::numeric_limits<std::uintmax_t>::max()) < -x)
 					? SPROUT_MATH_THROW_LARGE_FLOAT_ROUNDING(std::runtime_error("floor: large float rounding."), x)
 				: static_cast<FloatType>(sprout::math::detail::floor_impl(static_cast<typename sprout::math::detail::float_compute<FloatType>::type>(x)))
 #endif
