+++ include/sprout/math/round.hpp
@@ -74,7 +74,8 @@ namespace sprout {
 				: sprout::math::detail::builtin_round(x)
 #else
 				: x == 0 ? x
-				: sprout::numeric_limits<std::uintmax_t>::max() < x || sprout::numeric_limits<std::uintmax_t>::max() < -x
+                : (static_cast<FloatType>(sprout::numeric_limits<std::uintmax_t>::max()) <  x ||
+                   static_cast<FloatType>(sprout::numeric_limits<std::uintmax_t>::max()) < -x)
 					? SPROUT_MATH_THROW_LARGE_FLOAT_ROUNDING(std::runtime_error("round: large float rounding."), x)
 				: static_cast<FloatType>(sprout::math::detail::round_impl(static_cast<typename sprout::math::detail::float_compute<FloatType>::type>(x)))
 #endif
